requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
akshare>=1.12.0
tushare>=1.3.0
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
tenacity>=8.2.0
